+++ src/EventHub/EventHub.Autorest/custom/Set-AzEventHubNamespace.ps1
@@ -110,6 +110,11 @@ function Set-AzEventHubNamespace{
         [System.Int32]
         ${SkuCapacity},
 
+        [Parameter(HelpMessage = "The IP address type for the namespace. Determines whether the namespace supports IPv4 only or both IPv4 and IPv6.")]
+        [Microsoft.Azure.PowerShell.Cmdlets.EventHub.Category('Body')]
+        [System.String]
+        ${IPAddressType},
+
         [Parameter(HelpMessage = "The maximum acceptable lag for data replication operations from the primary replica to a quorum of secondary replicas. When the lag exceeds the configured amount, operations on the primary replica will be failed. The allowed values are 0 and 5 minutes to 1 day.")]
         [Microsoft.Azure.PowerShell.Cmdlets.EventHub.Category('Body')]
         [System.Int64]
@@ -197,6 +202,7 @@ function Set-AzEventHubNamespace{
             $hasSkuCapacity = $PSBoundParameters.Remove('SkuCapacity')
             $hasTag = $PSBoundParameters.Remove('Tag')
             $hasAsJob = $PSBoundParameters.Remove('AsJob')
+            $hasIPAddressType = $PSBoundParameters.Remove('IPAddressType')
             $hasGeoDataReplicationLocation = $PSBoundParameters.Remove('GeoDataReplicationLocation')
             $hasGeoDataReplicationMaxReplicationLagDurationInSecond = $PSBoundParameters.Remove('GeoDataReplicationMaxReplicationLagDurationInSecond')
             $null = $PSBoundParameters.Remove('WhatIf')
@@ -251,6 +257,9 @@ function Set-AzEventHubNamespace{
             if ($hasMinimumTlsVersion) {
                 $eventHubNamespace.MinimumTlsVersion = $MinimumTlsVersion
             }
+            if($hasIPAddressType){
+                   $eventHubNamespace.IPAddressType = $IPAddressType
+               }
             if ($hasPublicNetworkAccess) {
                 $eventHubNamespace.PublicNetworkAccess = $PublicNetworkAccess
             }